backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-genai==1.5.0
python-multipart==0.0.12
pydantic==2.9.0
